WebGPU acceleration for Merkle tree construction (#1)
* add feature to debug print in Merkle construction

We just print the size of each Merkle tree that is constructed and
when the construction is done.

This is like a 0-th order benchmark (*at least* 50% of the time is
spent constructing Merkle trees when running `prove_singles`, I think
more). But more importantly we now know the actual sizes for which we
construct Merkle trees!

* add scaffolding for WebGPU acceleration for Merkle trees

* get Merkle tree code compiling and "running"

Initial steps that sort of do something for some of the Merkle trees
part of `prove_singles`.

* update the WGSL code

GPT-5 made some random changes in there....

* work towards making things async

work towards making more calls in the plonky2 callstack from INTMAX
support Merkle trees on GPU by choosing the async route

* add standalone plonky2 Merkle tree GPU WASM "test case"

* get initial GPU merkle tree constr partially working

* add small test vectors for Merkle tree GPU test

* fix storage buffer / bind group layout for Merkle tree GPU

We still don't use uniform buffers. Storage so far. GPT-5 just assumed
they were storage already.

Now it actually works

* start towards hashing input vectors on GPU

* add timing for standalone Merkle tree construction test

* use GPU for hashing of inputs into leaf digests

* implement timing for Merkle GPU setup, execution, postprocessing

* implement pass to convert canonical to Montgomery rep

All data is now converted from canonical to Montgomery representation
with one kernel call. Much faster than doing it on the CPU with
terrible codegen.

* add WGSL files for Canonical -> Montgomery, Montgomery -> Canonical

* fix timing when running in web worker using `performance` API

* ask for 4GB allocations

* update WGSL code to use correct thread ID

* add conversion to canonical when writing to `cap` & non power 2 in hash

* handle Montgomery -> canonical conversion on GPU as well

* add debugging helper for FRI oracle prove

* extend starky prover with `async` support

* use async approach in plonk prover for partial products

* changes towards debugging deadlock in phase 4

* debug deadlock happening

the most important change is skipping all trees <= 4096 leaves

* add timing for NTT / INTT code

* add CPU vs GPU comparison for Merkle trees to standalone Plonky2 ex

This now constructs a Merkle tree on CPU and on GPU and compares them.

I've tried the new 8192 leaves tree and the large one. Both yield
identical results based on this code.

* extend GPU vs CPU test file with binary file parser

For ~500k leaves Merkle tree, where storing those as a string written
to the console doesn't work anymore.

* fix workgroup layout for all kernel launches

This fixes the proof verification failed bug in `prove_singles`. The
issue was that the canonical to Montgomery conversion step used way
too many workgroups (64 times as many), resulting in somehow broken
data. Not 100% sure what the issue was (in theory the old code should
have just launched 64 times too many kernels, so only 1/64 kernels
should have done real work).

* extend md file about WASM test for CPU vs GPU & binary data files

* further changes for async in `prove_test`

* fix runtime panic when running `prove_singles` twice in a row

* directly transform data into transposed words & drop canonical words

This avoids additional memory overhead. In `prove_test` we ran into an
unnamed runtime panic. This fixes that!

* progress on getting `prove_test` running on GPU

* explicitly annotate WGSL code with unrestricted pointer parameters extension

* change max storage buffer required to 1GB

* changes following WebGPU_readback_fixes.md

The only good thing about these changes (aside from all the annoying
syntax fixes that popped up for some reason) is that now the time in
the timing is associated with the kernel execution and not the very
first readback call.

* add Merkle tree construction benchmark running on native

* add GPU path to merkle benchmark

NOTE: On native the GPU code is not fully hooked up and so doesn't work!

* disable most logging unless `gpu_merkle_logging` feature used

* implement transpose & to montgomery on GPU in one kernel

- avoids duplicating the input leaves
- faster than CPU variant

NOTE: Currently the code deadlocks in the send chunk data function!

* fix bindings, buffer usage and `send_chunk_data` while loop

* fix workgroup size in transpose to mont kernel

* update Poseidon1, Merkle shaders to use uint math

* hide more logging behind new feature flag in merkle gpu code

* remove dead code

* more profiling code

* simplify and clean up merkle_tree_gpu.rs

* remove old debug output

* perform single read back and split that after

Instead of doing three for nodes, cap, leaf

* remove full readback feature from Cargo toml

* use lazy logger to avoid formatting strings unless logging used

While the formatting is not going to make a *huge* difference, there
is no point in doing string formatting when no message is ever going
to be printed. Minor performance opt

* improve subtree reconsrtuction on CPU, avoid recursive impl

Instead of a recursive subtree reconstruction, we now have an
iterative approach

* disable `wasm_test` if Plonky2 not what we build (via feature flag)

* update README

* remove `wasm_test` debugging / bench / testing helper

Without test vectors it's useless. Feel free to recover from the git
history and provide your own test vectors (or ping me @Vindaar for
larger inputs).

* remove additional file & Merkle tree CPU benchmark

---------

Co-authored-by: Morgan Thomas <morgan.a.s.thomas@gmail.com>